Update README.md
As a total newbie to excavator, I was stumped for a long time and could not make it work. What made it more confusing was the fact that there are various old excavator forks floating around which are configured the traditional way, via command line parameters. But this excavator version is different - it requires that all configuration is passed to it via JSON or API.

See this related issue: https://github.com/nicehash/excavator/issues/96

Eventually I ran excavator under `strace` and that's when I saw it was doing nothing after initialization. It finally dawned on me that this is really different, and the configuration must be somewhere else.

So I made this comment on README.md for folks like me who are confused by the unusual way that excavator is configured.